Adding documents to symfony structure
