+++ src/AppBundle/Document/Ticket.php
@@ -1,12 +1,15 @@
 <?php
 
-namespace Conticket\Document;
+namespace AppBundle\Document;
 
 use Doctrine\ODM\MongoDB\Mapping\Annotations as ODM;
 
 /** @ODM\EmbeddedDocument */
-final class Coupon
+final class Ticket
 {
+    const ACTIVE   = 'active';
+    const INACTIVE = 'inactive';
+
     /** @ODM\Id */
     private $id;
 
@@ -16,36 +19,41 @@ final class Coupon
     /** @ODM\String */
     private $description;
 
-    /** @ODM\String */
-    private $code;
+    /** @ODM\Int */
+    private $quantity;
 
     /** @ODM\Float */
     private $value;
 
-    /** @ODM\Int */
-    private $quantity;
+    /** @ODM\Date */
+    private $start;
 
     /** @ODM\Date */
-    private $expire;
+    private $end;
+
+    /** @ODM\String */
+    private $status;
 
     /**
      * Constructor
      *
      * @param string    $name
      * @param string    $description
-     * @param string    $code
-     * @param float     $value
      * @param int       $quantity
-     * @param \DateTime $expire
+     * @param float     $value
+     * @param \DateTime $start
+     * @param \DateTime $end
+     * @param string    $status
      */
-    public function __construct($name, $description, $code, $value, $quantity, \DateTime $expire)
+    public function __construct($name, $description, $quantity, $value, \DateTime $start, \DateTime $end, $status = self::ACTIVE)
     {
         $this->name        = $name;
         $this->description = $description;
-        $this->code        = $code;
-        $this->value       = (float) $value;
         $this->quantity    = (int) $quantity;
-        $this->expire      = $expire;
+        $this->value       = (float) $value;
+        $this->start       = $start;
+        $this->end         = $end;
+        $this->status      = $status;
     }
 
     public function getName()
@@ -58,9 +66,14 @@ final class Coupon
         return $this->description;
     }
 
-    public function getCode()
+    public function getStart()
     {
-        return $this->code;
+        return $this->start;
+    }
+
+    public function getEnd()
+    {
+        return $this->end;
     }
 
     public function getQuantity()
@@ -73,8 +86,8 @@ final class Coupon
         return $this->value;
     }
 
-    public function getExpire()
+    public function getStatus()
     {
-        return $this->expire;
+        return $this->status;
     }
 }
