+++ tests/ConticketFunctionalTests/AssertDocumentsShouldBeCreatedAndPersisted.phpt
@@ -10,7 +10,7 @@ use Doctrine\ODM\MongoDB\Configuration;
 use Doctrine\ODM\MongoDB\DocumentManager;
 use Doctrine\ODM\MongoDB\Mapping\Driver\AnnotationDriver;
 
-use Conticket\Document;
+use AppBundle\Document;
 
 AnnotationDriver::registerAnnotationClasses();
